+++ .coveragerc
@@ -6,9 +6,6 @@ source =
     spiel/
     tests/
 
-omit =
-    spiel/demo/*
-
 [report]
 
 skip_empty = True
