+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "spiel"
-version = "0.4.6"
+version = "0.5.0"
 description = "A framework for building and presenting richly-styled presentations in your terminal using Python."
 readme="README.md"
 homepage="https://github.com/JoshKarpel/spiel"
@@ -51,15 +51,14 @@ pre-commit = ">=3"
 pytest = ">=7"
 pytest-cov = ">=3"
 pytest-xdist = ">=3"
-mypy = ">=0.991"
-pytest-mypy = ">=0.10"
+pytest-asyncio = ">=0.20"
 pytest-mock = ">=3"
 hypothesis = ">=6"
-textual = {extras = ["dev"], version = "==0.9.1"}
+mypy = ">=0.991"
 mkdocs = ">=1.4"
 mkdocs-material = ">=9"
 mkdocstrings = {extras = ["python"], version = ">=0.19.0"}
-pytest-asyncio = ">=0.20"
+textual = {extras = ["dev"], version = "==0.9.1"}
 
 [tool.poetry.scripts]
 spiel = 'spiel.cli:cli'
